Add a separate file for the DOM manipulation
